+++ app/Models/Reconciliation.php
@@ -32,12 +32,13 @@ use RuntimeException;
  */
 class Reconciliation extends Model
 {
+    /** @use HasFactory<\Database\Factories\ReconciliationFactory> */
     use HasFactory;
 
     /**
      * The attributes that are mass assignable.
      *
-     * @var array<int, string>
+     * @var list<string>
      */
     protected $fillable = [
         'company_id',
@@ -66,7 +67,7 @@ class Reconciliation extends Model
         // Automatically set the reconciled_by_user_id and reconciled_at when creating
         static::creating(function (Reconciliation $reconciliation) {
             if (Auth::check()) {
-                $reconciliation->reconciled_by_user_id = Auth::id();
+                $reconciliation->reconciled_by_user_id = (int) Auth::id();
             }
             $reconciliation->reconciled_at = now();
         });
@@ -98,6 +99,9 @@ class Reconciliation extends Model
     /**
      * Get the company that owns this reconciliation.
      */
+    /**
+     * @return BelongsTo<Company, static>
+     */
     public function company(): BelongsTo
     {
         return $this->belongsTo(Company::class);
@@ -106,6 +110,9 @@ class Reconciliation extends Model
     /**
      * Get the user who performed this reconciliation.
      */
+    /**
+     * @return BelongsTo<User, static>
+     */
     public function reconciledBy(): BelongsTo
     {
         return $this->belongsTo(User::class, 'reconciled_by_user_id');
@@ -113,6 +120,9 @@ class Reconciliation extends Model
 
     /**
      * Get the journal entry lines that are part of this reconciliation.
+     */
+    /**
+     * @return BelongsToMany<JournalEntryLine, static>
      */
     public function journalEntryLines(): BelongsToMany
     {
